Adding Scripts/updated scripts
Updated the ETOPO script to save the ETOPO_Downsampled image in the Assets folder, displayed in the notebook. LWE_Thickness_animation outputs both the grace_lwe_animation and LWE Gravity Change gifs. The Satellite_Orbit is for the sphere and satellite animation stores in the Assets aswell. I hereby promise that I shall add comments.