Add files via upload
!!! This code is an edited version of the code and explanation found here: https://github.com/jacopok/anomalies-visualization

The username of the person who uploaded this is jacopok

The repository is called anomalies_visualization

I just edited it to make it a little easier on my eyes, and able to work from one file. A lot of this project is so I can learn what I think is needed for my project. I learned a lot from jacopok's repository...so go look at it.