Scripts for Satellite animation and groundtracks
These scripts are for the Satellite animation and the groundtracks figure. They are based on satellite data, and they assume that you already have TLE files ready.

For the Satellite_Orbit script, you'll need to have a destination set to your GRACE-FO and ICESat-2 TLEs. They should be .txt files. You'll need to define them as "grace" and "icesat". Or...whatever you want, you'll just need to go in and rename everything if you want to use something different.

For the TLE_Handler...I realised you could just grab both from the folder they're in. You don't need to do it that way, I'm just experimenting with what I can do.

These make use of the Skyfield package for python. It's uh...an amazing package.