Script that downsamples the geoid resolution
This is the script that is responsible for the down-sampled Geoid topography figure.

I'd be careful about running this. I tried to make this use the least amount of memory as possible, but that might not stop your computer from going super saiyan.

Head my comments, do not try to run that topography dataset unless its low resolution.